Cleanup; drag&drop is working now
