--- src/tests/unit/index.html
@@ -1,19 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <title></title>
-    <script src="../../bower_components/qunit/qunit/qunit.js"></script>
-    <script src="/testem.js"></script>
-    <script src="../../dist/pulltorefresh.js"></script>
-    <script src="tests.js"></script>
-
-    <link rel="stylesheet" href="../../bower_components/qunit/qunit/qunit.css">
-</head>
-<body>
-    <div id="qunit"></div>
-    <div id="qunit-fixture"></div>
-
-    <div id="ptr-test-container">
-        <div id="ptr-trigger-element"></div>
-    </div>
-</body>
