+++ tests/unit/tests.js
@@ -1,62 +1,56 @@
-(function () {
-  
+/* global QUnit, PullToRefresh */
 
-  var test = QUnit.test;
-  var module$1 = QUnit.module;
+const { test, module } = QUnit;
 
-  var _ = PullToRefresh._;
-  var $ = function (selector) { return document.querySelector(selector); };
-  var $$ = function (selector) { return document.querySelectorAll(selector); };
+const _ = PullToRefresh._;
+const $ = selector => document.querySelector(selector);
+const $$ = selector => document.querySelectorAll(selector);
 
-  module$1('Setup & Teardown');
-  test('init() & destroy()', function (assert) {
-    assert.equal(typeof PullToRefresh.init, 'function', 'Init is a function');
+module('Setup & Teardown');
+test('init() & destroy()', assert => {
+  assert.equal(typeof PullToRefresh.init, 'function', 'Init is a function');
 
-    var x = PullToRefresh.init({ mainElement: '#ptr-trigger-element', triggerElement: '#ptr-trigger-element' });
+  const x = PullToRefresh.init({ mainElement: '#ptr-trigger-element', triggerElement: '#ptr-trigger-element' });
 
-    assert.equal(typeof x.destroy, 'function', 'returns a destroy function');
+  assert.equal(typeof x.destroy, 'function', 'returns a destroy function');
 
-    // attach DOM & events manually
-    _.setupDOM(x);
+  // attach DOM & events manually
+  _.setupDOM(x);
 
-    assert.ok($('.ptr--ptr'), 'it adds the ptr element');
-    assert.equal($$('#pull-to-refresh-js-style').length, 1, 'it adds the style element');
+  assert.ok($('.ptr--ptr'), 'it adds the ptr element');
+  assert.equal($$('#pull-to-refresh-js-style').length, 1, 'it adds the style element');
 
-    x.destroy();
+  x.destroy();
 
-    _.onReset(x);
+  _.onReset(x);
 
-    var done = assert.async();
+  const done = assert.async();
 
-    setTimeout(function () {
-      assert.equal($$('.ptr--ptr').length, 0, 'destroy removes the ptr element');
-      assert.equal($$('#pull-to-refresh-js-style').length, 0, 'destroy removes the style element');
+  setTimeout(() => {
+    assert.equal(document.querySelectorAll('.ptr--ptr').length, 0, 'destroy removes the ptr element');
 
-      done();
-    }, x.refreshTimeout + 1);
-  });
+    done();
+  }, x.refreshTimeout + 1);
+});
 
-  test('Ensure Init is idempotent', function (assert) {
-    var x = PullToRefresh.init({ mainElement: '#ptr-trigger-element', triggerElement: '#ptr-trigger-element' });
-    var y = PullToRefresh.init({ mainElement: '#ptr-trigger-element', triggerElement: '#ptr-trigger-element' });
+test('Ensure Init is idempotent', assert => {
+  const x = PullToRefresh.init({ mainElement: '#ptr-trigger-element', triggerElement: '#ptr-trigger-element' });
+  const y = PullToRefresh.init({ mainElement: '#ptr-trigger-element', triggerElement: '#ptr-trigger-element' });
 
-    _.setupDOM(x);
-    _.setupDOM(y);
+  _.setupDOM(x);
+  _.setupDOM(y);
 
-    assert.equal($$('.ptr--ptr').length, 2, 'add two ptr elements');
-    assert.equal($$('#pull-to-refresh-js-style').length, 1, 'adds the style element once');
+  assert.equal($$('.ptr--ptr').length, 2, 'add two ptr elements');
+  assert.equal($$('#pull-to-refresh-js-style').length, 1, 'adds the style element once');
 
-    var done = assert.async();
+  const done = assert.async();
 
-    _.onReset(x);
-    _.onReset(y);
+  _.onReset(x);
+  _.onReset(y);
 
-    setTimeout(function () {
-      assert.equal($$('.ptr--ptr').length, 0, 'remove added ptr elements');
-      assert.equal($$('#pull-to-refresh-js-style').length, 0, 'remove style element');
+  setTimeout(() => {
+    assert.equal(document.querySelectorAll('.ptr--ptr').length, 0, 'remove added ptr elements');
 
-      done();
-    }, x.refreshTimeout + 2);
-  });
-
-}());
+    done();
+  }, x.refreshTimeout + 2);
+});
